refactor: move `src/pipx.mts` to `src/pipx/install.mts`
